My Final Year Project - Vision-Based Emotion Detection of Gameplayers
This is my final year project, which is a program which detects the emotions of a person's face. The program was designed for this project and its experimentation step, which was to detect the emotions of the participants whilst they played a game of my choosing over three different time periods of 20 minutes, 40 minutes and 60 minutes. The results collected were analysed and a subsequent conclusion was made, which was found to be seen that people enjoy shorted gameplays when compared to the longer ones. The project's report and presentation will also be included in a later commit.